+++ Assets/AutoTranslate/Scripts/GUI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 30af6c66dd567224d93b24570b0dd864
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
