+++ Assets/AutoTranslate/Scripts/Helpers/GameObjectHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4a9fef2ca9d36d644964f51c99b47398
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
