+++ Assets/AutoTranslate/Scripts/InterfaceTranslate.asmdef
@@ -1,6 +1,10 @@
 {
     "name": "InterfaceTranslate",
-    "references": [],
+    "rootNamespace": "",
+    "references": [
+        "GUID:eec0964c48f6f4e40bc3ec2257ccf8c5",
+        "GUID:2e77701fe0ceda44fb763f03ee704c37"
+    ],
     "includePlatforms": [
         "Editor"
     ],
