+++ Assets/AutoTranslate/Scripts/LocalizeDropdownEvent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e4bfe6b2382c2464c91cc16b81385079
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
