+++ Assets/AutoTranslate/Scripts/SimpleLocalization.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 22ffe6ddda1fac946a878372121a6a32
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
