+++ Assets/AutoTranslate/Scripts/SimpleLocalization/LocalizeStringEventExtension.cs.meta
