+++ Assets/AutoTranslate/Scripts/SimpleLocalization/SharedTableDataExtension.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8e1e716478e44384c80f809622e7c911
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
