+++ Assets/AutoTranslate/Scripts/SimpleLocalization/SimpleInterfaceLocalization.cs.meta
