+++ Assets/AutoTranslate/Scripts/SimpleLocalization/SimpleInterfaceStringTable.cs.meta
